...
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@446 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ generic/keydisp.c
@@ -540,7 +540,7 @@ static void setpsghdl(BYTE items) {
 	}
 }
 
-void keydisp_setfmboard(BYTE b) {
+void keydisp_setfmboard(UINT b) {
 
 	keydisp.keymax = 0;
 	keydisp.fmmax = 0;
